merge from origne master

+++ src/server/game/Server/Packets/ArtifactPackets.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2008-2016 TrinityCore <http://www.trinitycore.org/>
+ * Copyright (C) 2008-2017 TrinityCore <http://www.trinitycore.org/>
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
@@ -97,7 +97,7 @@ namespace WorldPackets
             WorldPacket const* Write() override;
 
             ObjectGuid ArtifactGUID;
-            int32 Amount = 0;
+            uint64 Amount = 0;
         };
     }
 }
